+++ src/dials/algorithms/simulation/reciprocal_space_helpers.h
@@ -52,7 +52,7 @@ namespace dials { namespace algorithms {
     const int6& bbox,
     std::size_t I,
     af::ref<double, af::c_grid<3> > shoebox,
-    const af::const_ref<int, af::c_grid<3> >& mask) {
+    const af::const_ref<uint8_t, af::c_grid<3> >& mask) {
     vec3<double> s0 = beam.get_s0();
     vec3<double> m2 = goniometer.get_rotation_axis();
 
@@ -118,7 +118,7 @@ namespace dials { namespace algorithms {
     double phi,
     const int6& bbox,
     std::size_t I,
-    const af::const_ref<int, af::c_grid<3> >& mask) {
+    const af::const_ref<uint8_t, af::c_grid<3> >& mask) {
     vec3<double> s0 = beam.get_s0();
     vec3<double> m2 = goniometer.get_rotation_axis();
